+++ src/components/auth/components/Login.jsx
@@ -1,41 +1,12 @@
 "use client";
 import { Button, Input } from "@nextui-org/react";
 import Link from "next/link";
-import { API_URL } from "@/config/apiUrl";
-import { useState } from "react";
+import { useLogin } from "../hooks/useLogin";
 
 export default function Login() {
-  // const { handleChange, handleLogin, loading } = useRegister();
-  const [loading, setLoading] = useState(false);
-  const [loginData, setLoginData] = useState({
-    email: "",
-    password:""
-  })
-  function handleChange(e){
-    const { name, value } = e.target;
-    setLoginData({ ...loginData, [name]: value });
-    
-  }
-  async function handleSubmitLogin() {
-    setLoading(true);
-    const { email, password } = loginData;
-    const res = await fetch(`${API_URL}/api/auth/login`,{
-      method: "POST",
-      headers: {
-        "Content-Type":"application/json"
-      },
-      body: JSON.stringify({email, password})
-    })
-    const data = await res.json();
-    if (!data) {
-      setLoading(false);
-      console.log("error");
-      return
-    }
-    setLoading(false);
-    console.log(data);
-    
-  }
+  
+  const { handleChange, handleSubmitLogin, loading } = useLogin();
+  
   return (
     <main className="space-y-3">
       <h3>Welcome Back!</h3>
